refactor(thrid-party): rename section (#69)
