+++ src/Contracts/Expr/CorrespondingAssertInterface.php
@@ -7,7 +7,7 @@ namespace StructuraPhp\Structura\Contracts\Expr;
 use Closure;
 use StructuraPhp\Structura\ValueObjects\ClassDescription;
 
-interface OtherAssertInterface
+interface CorrespondingAssertInterface
 {
     /**
      * @param Closure(ClassDescription): string $closure
@@ -43,22 +43,4 @@ interface OtherAssertInterface
      * @param Closure(ClassDescription): string $closure
      */
     public function toHaveCorrespondingTrait(Closure $closure, string $message = ''): self;
-
-    /**
-     * @param array<int,string>|string $patterns class names or regular expression patterns to
-     *                                           be matched with namespaces
-     */
-    public function toBeInOneOfTheNamespaces(
-        array|string $patterns,
-        string $message = '',
-    ): self;
-
-    /**
-     * @param array<int,string>|string $patterns class names or regular expression patterns not
-     *                                           to be matched with namespaces
-     */
-    public function notToBeInOneOfTheNamespaces(
-        array|string $patterns,
-        string $message = '',
-    ): self;
 }
