+++ src/Contracts/Expr/ThirdPartyAssertInterface.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace StructuraPhp\Structura\Contracts\Expr;
+
+interface ThirdPartyAssertInterface
+{
+    /**
+     * @param array<int,string>|string $patterns class names or regular expression patterns to
+     *                                           be matched with namespaces
+     */
+    public function toBeInOneOfTheNamespaces(
+        array|string $patterns,
+        string $message = '',
+    ): self;
+
+    /**
+     * @param array<int,string>|string $patterns class names or regular expression patterns not
+     *                                           to be matched with namespaces
+     */
+    public function notToBeInOneOfTheNamespaces(
+        array|string $patterns,
+        string $message = '',
+    ): self;
+}
