+++ tests/Unit/Services/AnalyseServiceTest.php
@@ -90,7 +90,7 @@ final class AnalyseServiceTest extends TestCase
              & to extend <promote>BadMethodCallException</promote>
 
         <pass> PASS </pass> Asserts architecture rules in StructuraPhp\Structura\Tests\Feature\TestVoid
-        137 classe(s) from
+        139 classe(s) from
          - dirs
         That
         Should
